edit

+++ src/services/todo/todo.class.js
@@ -23,7 +23,7 @@ export class TodoService {
     try {
       const { userId } = _params.user;
 
-      return await this.Model.findOne({ _id: id ,user:userId});
+      return await this.Model.findOne({ _id: id, user: userId });
     } catch (error) {
       throw new Error('Error occurred while getting the TODO.');
     }
@@ -41,6 +41,13 @@ export class TodoService {
 
   async update(id, data, _params) {
     try {
+      const { userId } = _params.user;
+      const todo = await this.Model.findOne({ _id: id, user: userId });
+
+      if (!todo) {
+        throw new Error('TODO not found or user is not the owner.');
+      }
+
       return await this.Model.findByIdAndUpdate(id, data, { new: true });
     } catch (error) {
       throw new Error('Error occurred while updating the TODO.');
@@ -49,6 +56,13 @@ export class TodoService {
 
   async patch(id, data, _params) {
     try {
+      const { userId } = _params.user;
+      const todo = await this.Model.findOne({ _id: id, user: userId });
+
+      if (!todo) {
+        throw new Error('TODO not found or user is not the owner.');
+      }
+
       return await this.Model.findOneAndUpdate({ _id: id }, data, { new: true });
     } catch (error) {
       throw new Error('Error occurred while patching the TODO.');
@@ -56,7 +70,15 @@ export class TodoService {
   }
 
   async remove(id, _params) {
+    console.log(_params)
     try {
+      const { userId } = _params.user;
+      const todo = await this.Model.findOne({ _id: id, user: userId });
+
+      if (!todo) {
+        throw new Error('TODO not found or user is not the owner.');
+      }
+
       return await this.Model.findByIdAndRemove(id);
     } catch (error) {
       throw new Error('Error occurred while removing the TODO.');
